requirements.txt
attrs==22.1.0
bio==1.4.0
biopython==1.79
biothings-client==0.2.6
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.34.4
idna==3.3
iniconfig==1.1.1
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.5.3
munch==2.5.0
mygene==3.2.2
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.11
yapf==0.32.0
